Methods Update (Electrical Losses, Turbine Gross, Gap Analysis) (#38)
This pull request synthesizes changes to three fundamental methods for OpenOA:

-Electrical loss calculation
-Turbine gross calculation
-Gap Analysis calculation

For electrical losses, the main change was introducing a procedure for meter data with a monthly resolution that leveraged a "correction threshold." This correction threshold is dictated by the user and stipulates what percentage of full reporting is required for data to be included.

For turbine gross, there are two main changes. The first involves moving/re-organizing the filtering to be more generalizable to different projects. This can be seen through a large number of new parameters. The second introduces imputing into the turbine gross calculation as a protection from low reporting.

For the gap analysis, a small plotting adaption was made so that a plant/review-specific title can be included in the title of the plot.



* Add drop nan line in turb long term gross energy to fix model fitting

* Add extra window filter to turbine_long_term_gross_energy method

* Fix Electrical loss local merge conflict take 2

* Introduce perc column for monthly scada to improve monthly electric loss estimates

* Add scada sum as instance variable

* Add more comments and line breaks for clarity

* Introduce leap year correction

* Revert turbine_long_term_gross to current develop state

* Add parametrization for correction threshold

* Use built-in datetime functionality to handle reporting filtering

* Update electrical losses calc

* Add diff array as instance variable

* Update turbine_long_gross_energy for imputing in anticipation of pull request

* Update electrical loss and turbine gross tests with more parameters

* Remove partial array of turbines