Feature/cubico projects (#222)
* Update .gitignore

Ignore kelmarsh and penmanshiel data folders

* Project Cubico created

This allows the 02_plant_aep_analysis_Cubico to be run for both Kelmarsh and Penmanshiel wind farms

* Update to include MERRA2 data

Still need to test, limit to region, and read into the analysis as a reanalysis dataset

* Updated get_merra2 for nearest 9 nodes

Now only downloads nearest 9 nodes to the coordinates, saving on storage

* Reanalysis validator updated to accept any string

Previously this was limited specifically to "merra2", "ncep2", "erai", "era5," but other names and sources are likely to be used

* Monthly MERRA2 is now available in project_Cubico

Both ERA5 and MERRA2 monthly time series are now automatically acquired for the long-term analysis

* Additional requirements for reanalysis downloads

Adding the requirements for downloading the monthly ERA5 and MERRA2 data

* Separated downloading for project data into utils/downloader

The functions to download data from the internet and from CDS and GES DISC are now in the utils/downloader.py file, so that they can be used outside of the project_Cubico.py file

* Improved comments and formatting

Tested with both Kelmarsh and Penmanshiel - appears to work fine

* converted module docstring to markdown-compatible commentary

* add static typing to example script

* resort imports and initializations for consistency

* basic reformatting for google docstring style

* replace os usage with Path

* update for efficient chaining

* Path compatibility

* f-strings for string concatenation

* updating imports and initializations

* reformat list for proper display in docs

* add static typing to downloader

* f-string and Path refactor

* drop now obselete os import

* add Path option and add a missing f-string

* missed updates for f-strings and Path

* update pre-commit hook versions for failed docs building

* fix typo in update to f-string

* fix bad f-string

* update ci workflow to test if version change fixes bug

* updates to pass pre-commit hook

* Converting double slash to single slash

// -> /

* Added Raises NameError to docstring

Done for the download_asset_data function

* Added Raises NameError to docstring

This is for the get_era5 function, also changed error message to f-string

* Standardising variable name to snake_case

SCADA when in variables is now lowercase, as used elsewhere in the codebase, a couple of other variables were also updated (zipFiles and usecolumns). Some other typos were also corrected. ERA5 and MERRA2 are no longer hyphenated anywhere

* reformatting for google docstring style

* replaced print output with tqdm progress bar

* Added traceback to IO error

* Added raised exceptions to download_file function

HTTPError: If unable to access url.
Exception: If the request failed for another reason.

* Rename file variable to f

* Added start_year and end_year as keyword arguments

Applied to get_era5 and get_merra2 to allow users to select the range of years, with a simple error check that the years make sense

* merra/era functions appended with "monthly"

To make it clear that the data downloaded is monthly resolution

* merra and era function docstrings updated

Listing variables and location of saved files in the description

* get_era5_monthly now has an update dict

A base dict is used in the ERA5 request and only month and year updated in the loop

* Improving reanalysis downloaders

Only downloads requested and available months

* extract_all_data function removed

This was previously used to unzip all files in the specified directory

* Docstring improved by removing empty Returns

Where now output is provided in the return, any saved files etc. should be detailed in the main docstring

* Docstring improved by removing empty Returns

* Updated to the new 61400-25 tags

* Merge in latest Dev 3 changes

Update includes the required new IEC naming conventions

* Fixing IEC tags

* Set MERRA2 function to always save a csv

* Formatting comments

* Fix issue with empty rows in asset data

* Updates to the Cubico notebook

* ensure there is a way to install the required packages for era5 usage

* fixing the .* issue and adding in era5 requirements

* add netcdf4 for merra2

* add a little more info on data downloading package requirements

* changed cubico example notebook to 02c; minor notebook commentary updates

* minor docstring fix in project_Cubico

* adding Cubico contributions to changelog

* add missing test update

* update examples following new dependencies and feature add

---------

Co-authored-by: Rob Hammond <13874373+RHammond2@users.noreply.github.com>
Co-authored-by: ejsimley <eric.simley@nrel.gov>